requirements/base.txt
boto==2.49.0
dj-rest-auth==2.1.3
django-allauth==0.44.0
Django==2.2.16
django-cors-headers==3.1.0
django-filter==2.2.0
django-import-export==1.2.0
django-modeltranslation==0.16.1
django-storages==1.7.1
django-unique-upload==0.2.1
djangorestframework==3.12.2
djangorestframework-simplejwt
drf-yasg==1.20.0
mysqlclient==1.4.4
Pillow==8.1.0
python-decouple==3.4
django-cities-light==3.8.1 # https://github.com/yourlabs/django-cities-light
django-smart-selects==1.5.9
django-solo==1.1.5
firebase_admin